refactor: Move Auth Login to Member
