+++ src/main/kotlin/ecommerce/controller/member/MemberAuthController.kt
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 @RequestMapping("/api/auth")
-class AuthController(
+class MemberAuthController(
     val authService: AuthService,
 ) {
     @PostMapping("/signUp")
